+++ src/BraceCompletion/BraceMatchingTagger.cs
@@ -1,0 +1,227 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.VisualStudio.Text;
+using Microsoft.VisualStudio.Text.Editor;
+using Microsoft.VisualStudio.Text.Tagging;
+
+namespace MarkdownEditor
+{
+    public class BraceMatchingTagger : ITagger<TextMarkerTag>
+    {
+        ITextView View { get; set; }
+        ITextBuffer SourceBuffer { get; set; }
+        SnapshotPoint? CurrentChar { get; set; }
+        private readonly Dictionary<char, char> _braceList;
+
+        internal BraceMatchingTagger(ITextView view, ITextBuffer sourceBuffer)
+        {
+            //here the keys are the open braces, and the values are the close braces
+            _braceList = new Dictionary<char, char> { { '[', ']' }, { '(', ')' }, { '{', '}' } };
+            View = view;
+            SourceBuffer = sourceBuffer;
+            CurrentChar = null;
+
+            View.Caret.PositionChanged += CaretPositionChanged;
+            View.LayoutChanged += ViewLayoutChanged;
+        }
+
+        void ViewLayoutChanged(object sender, TextViewLayoutChangedEventArgs e)
+        {
+            if (e.NewSnapshot != e.OldSnapshot) //make sure that there has really been a change
+            {
+                UpdateAtCaretPosition(View.Caret.Position);
+            }
+        }
+
+        public IEnumerable<ITagSpan<TextMarkerTag>> GetTags(NormalizedSnapshotSpanCollection spans)
+        {
+            if (spans.Count == 0)   //there is no content in the buffer
+                yield break;
+
+            //don't do anything if the current SnapshotPoint is not initialized or at the end of the buffer
+            if (!CurrentChar.HasValue || CurrentChar.Value.Position >= CurrentChar.Value.Snapshot.Length)
+                yield break;
+
+            //hold on to a snapshot of the current character
+            SnapshotPoint currentChar = CurrentChar.Value;
+
+            //if the requested snapshot isn't the same as the one the brace is on, translate our spans to the expected snapshot
+            if (spans[0].Snapshot != currentChar.Snapshot)
+            {
+                currentChar = currentChar.TranslateTo(spans[0].Snapshot, PointTrackingMode.Positive);
+            }
+
+            char currentText;
+
+            try
+            {
+                //get the current char and the previous char
+                currentText = currentChar.GetChar();
+            }
+            catch (Exception)
+            {
+                yield break;
+            }
+
+            SnapshotPoint lastChar = currentChar == 0 ? currentChar : currentChar - 1; //if currentChar is 0 (beginning of buffer), don't move it back
+            char lastText = lastChar.GetChar();
+            var pairSpan = new SnapshotSpan();
+
+            if (_braceList.ContainsKey(currentText))   //the key is the open brace
+            {
+                _braceList.TryGetValue(currentText, out char closeChar);
+                if (FindMatchingCloseChar(currentChar, currentText, closeChar, View.TextViewLines.Count, out pairSpan) == true)
+                {
+                    yield return new TagSpan<TextMarkerTag>(new SnapshotSpan(currentChar, 1), new TextMarkerTag("MarkerFormatDefinition/HighlightWordFormatDefinition"));
+                    yield return new TagSpan<TextMarkerTag>(pairSpan, new TextMarkerTag("MarkerFormatDefinition/HighlightWordFormatDefinition"));
+                }
+            }
+            else if (_braceList.ContainsValue(lastText))    //the value is the close brace, which is the *previous* character
+            {
+                IEnumerable<char> open = from n in _braceList
+                                         where n.Value.Equals(lastText)
+                                         select n.Key;
+                if (FindMatchingOpenChar(lastChar, open.ElementAt(0), lastText, View.TextViewLines.Count, out pairSpan))
+                {
+                    yield return new TagSpan<TextMarkerTag>(new SnapshotSpan(lastChar, 1), new TextMarkerTag("MarkerFormatDefinition/HighlightWordFormatDefinition"));
+                    yield return new TagSpan<TextMarkerTag>(pairSpan, new TextMarkerTag("MarkerFormatDefinition/HighlightWordFormatDefinition"));
+                }
+            }
+        }
+
+        private static bool FindMatchingCloseChar(SnapshotPoint startPoint, char open, char close, int maxLines, out SnapshotSpan pairSpan)
+        {
+            if (startPoint.Snapshot.Length < 2)
+            {
+                pairSpan = new SnapshotSpan(startPoint.Snapshot, 0, startPoint.Snapshot.Length);
+                return false;
+            }
+
+            pairSpan = new SnapshotSpan(startPoint.Snapshot, 1, 1);
+            ITextSnapshotLine line = startPoint.GetContainingLine();
+            string lineText = line.GetText();
+            int lineNumber = line.LineNumber;
+            int offset = startPoint.Position - line.Start.Position + 1;
+
+            int stopLineNumber = startPoint.Snapshot.LineCount - 1;
+            if (maxLines > 0)
+                stopLineNumber = Math.Min(stopLineNumber, lineNumber + maxLines);
+
+            int openCount = 0;
+            while (true)
+            {
+                //walk the entire line
+                while (offset < line.Length)
+                {
+                    char currentChar = lineText[offset];
+                    if (currentChar == close) //found the close character
+                    {
+                        if (openCount > 0)
+                        {
+                            openCount--;
+                        }
+                        else    //found the matching close
+                        {
+                            pairSpan = new SnapshotSpan(startPoint.Snapshot, line.Start + offset, 1);
+                            return true;
+                        }
+                    }
+                    else if (currentChar == open) // this is another open
+                    {
+                        openCount++;
+                    }
+                    offset++;
+                }
+
+                //move on to the next line
+                if (++lineNumber > stopLineNumber)
+                    break;
+
+                line = line.Snapshot.GetLineFromLineNumber(lineNumber);
+                lineText = line.GetText();
+                offset = 0;
+            }
+
+            return false;
+        }
+
+        private static bool FindMatchingOpenChar(SnapshotPoint startPoint, char open, char close, int maxLines, out SnapshotSpan pairSpan)
+        {
+            pairSpan = new SnapshotSpan(startPoint, startPoint);
+
+            ITextSnapshotLine line = startPoint.GetContainingLine();
+
+            int lineNumber = line.LineNumber;
+            int offset = startPoint - line.Start - 1; //move the offset to the character before this one
+
+            //if the offset is negative, move to the previous line
+            if (offset < 0 && lineNumber > 0)
+            {
+                line = line.Snapshot.GetLineFromLineNumber(--lineNumber);
+                offset = line.Length - 1;
+            }
+
+            string lineText = line.GetText();
+
+            int stopLineNumber = 0;
+            if (maxLines > 0)
+                stopLineNumber = Math.Max(stopLineNumber, lineNumber - maxLines);
+
+            int closeCount = 0;
+
+            while (true)
+            {
+                // Walk the entire line
+                while (offset >= 0)
+                {
+                    char currentChar = lineText[offset];
+
+                    if (currentChar == open)
+                    {
+                        if (closeCount > 0)
+                        {
+                            closeCount--;
+                        }
+                        else // We've found the open character
+                        {
+                            pairSpan = new SnapshotSpan(line.Start + offset, 1); //we just want the character itself
+                            return true;
+                        }
+                    }
+                    else if (currentChar == close)
+                    {
+                        closeCount++;
+                    }
+                    offset--;
+                }
+
+                // Move to the previous line
+                if (--lineNumber < stopLineNumber)
+                    break;
+
+                line = line.Snapshot.GetLineFromLineNumber(lineNumber);
+                lineText = line.GetText();
+                offset = line.Length - 1;
+            }
+            return false;
+        }
+
+        void CaretPositionChanged(object sender, CaretPositionChangedEventArgs e)
+        {
+            UpdateAtCaretPosition(e.NewPosition);
+        }
+
+        void UpdateAtCaretPosition(CaretPosition caretPosition)
+        {
+            CurrentChar = caretPosition.Point.GetPoint(SourceBuffer, caretPosition.Affinity);
+
+            if (!CurrentChar.HasValue)
+                return;
+
+            TagsChanged?.Invoke(this, new SnapshotSpanEventArgs(new SnapshotSpan(SourceBuffer.CurrentSnapshot, 0, SourceBuffer.CurrentSnapshot.Length)));
+        }
+
+        public event EventHandler<SnapshotSpanEventArgs> TagsChanged;
+    }
+}