+++ src/BraceCompletion/BraceMatchingTaggerProvider.cs
@@ -1,0 +1,26 @@
+﻿using System.ComponentModel.Composition;
+using Microsoft.VisualStudio.Text;
+using Microsoft.VisualStudio.Text.Editor;
+using Microsoft.VisualStudio.Text.Tagging;
+using Microsoft.VisualStudio.Utilities;
+
+namespace MarkdownEditor
+{
+    [Export(typeof(IViewTaggerProvider))]
+    [ContentType(MarkdownLanguage.LanguageName)]
+    [TagType(typeof(TextMarkerTag))]
+    class BraceMatchingTaggerProvider : IViewTaggerProvider
+    {
+        public ITagger<T> CreateTagger<T>(ITextView textView, ITextBuffer buffer) where T : ITag
+        {
+            if (!MarkdownEditorPackage.Language.Preferences.EnableShowMatchingBrace)
+                return null;
+
+            //provide highlighting only on the top-level buffer
+            if (textView.TextBuffer != buffer)
+                return null;
+
+            return new BraceMatchingTagger(textView, buffer) as ITagger<T>;
+        }
+    }
+}